requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --strip-extras requirements.in
#
blinker==1.8.2
    # via flask
click==8.1.7
    # via flask
dnspython==2.6.1
    # via pymongo
flask==3.0.3
    # via
    #   -r requirements.in
    #   flask-cors
    #   flask-httpauth
flask-cors==4.0.1
    # via -r requirements.in
flask-httpauth==4.8.0
    # via -r requirements.in
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via flask
loguru==0.7.2
    # via -r requirements.in
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
mongoengine==0.28.2
    # via -r requirements.in
pillow==10.3.0
    # via -r requirements.in
pymongo==4.8.0
    # via mongoengine
werkzeug==3.0.3
    # via flask